+++ fiboa_cli/datasets/ch_be.py
@@ -1,0 +1,5 @@
+from .ch_base import CHBaseConverter
+
+
+class Converter(CHBaseConverter):
+    canton = "BE"